extract a trait
